fixed tests for flux lutman
